adds the portfolio dashboard option and compliance to the global menu
