+++ ui/components/multichain/account-list-menu/account-list-menu.js
@@ -3,12 +3,12 @@ import PropTypes from 'prop-types';
 import { useHistory } from 'react-router-dom';
 import Fuse from 'fuse.js';
 import { useDispatch, useSelector } from 'react-redux';
-import Box from '../../ui/box/box';
 import {
   IconName,
   ButtonLink,
   TextFieldSearch,
   Text,
+  Box,
 } from '../../component-library';
 import { AccountListItem, CreateAccount, ImportAccount } from '..';
 import {
@@ -19,12 +19,6 @@ import {
 import { useI18nContext } from '../../../hooks/useI18nContext';
 import { MetaMetricsContext } from '../../../contexts/metametrics';
 import Popover from '../../ui/popover';
-///: BEGIN:ONLY_INCLUDE_IN(build-mmi)
-import {
-  getMmiPortfolioEnabled,
-  getMmiPortfolioUrl,
-} from '../../../selectors/institutional/selectors';
-///: END:ONLY_INCLUDE_IN
 import {
   getSelectedAccount,
   getMetaMaskAccountsOrdered,
@@ -41,7 +35,6 @@ import {
   CONNECT_HARDWARE_ROUTE,
   ///: BEGIN:ONLY_INCLUDE_IN(build-mmi)
   CUSTODY_ACCOUNT_ROUTE,
-  COMPLIANCE_FEATURE_ROUTE,
   ///: END:ONLY_INCLUDE_IN
 } from '../../../helpers/constants/routes';
 import { getEnvironmentType } from '../../../../app/scripts/lib/util';
@@ -57,11 +50,6 @@ export const AccountListMenu = ({ onClose }) => {
   const history = useHistory();
   const dispatch = useDispatch();
   const inputRef = useRef();
-
-  ///: BEGIN:ONLY_INCLUDE_IN(build-mmi)
-  const mmiPortfolioUrl = useSelector(getMmiPortfolioUrl);
-  const mmiPortfolioEnabled = useSelector(getMmiPortfolioEnabled);
-  ///: END:ONLY_INCLUDE_IN
 
   const [searchQuery, setSearchQuery] = useState('');
   const [actionMode, setActionMode] = useState('');
@@ -198,6 +186,7 @@ export const AccountListMenu = ({ onClose }) => {
                 size={Size.SM}
                 startIconName={IconName.Add}
                 onClick={() => {
+                  dispatch(toggleAccountMenu());
                   trackEvent({
                     category: MetaMetricsEventCategory.Navigation,
                     event: MetaMetricsEventName.AccountAddSelected,
@@ -206,7 +195,7 @@ export const AccountListMenu = ({ onClose }) => {
                       location: 'Main Menu',
                     },
                   });
-                  setActionMode('add');
+                  history.push(NEW_ACCOUNT_ROUTE);
                 }}
                 data-testid="multichain-account-menu-add-account"
               >
@@ -218,6 +207,7 @@ export const AccountListMenu = ({ onClose }) => {
                 size={Size.SM}
                 startIconName={IconName.Import}
                 onClick={() => {
+                  dispatch(toggleAccountMenu());
                   trackEvent({
                     category: MetaMetricsEventCategory.Navigation,
                     event: MetaMetricsEventName.AccountAddSelected,
@@ -226,13 +216,13 @@ export const AccountListMenu = ({ onClose }) => {
                       location: 'Main Menu',
                     },
                   });
-                  setActionMode('import');
+                  history.push(IMPORT_ACCOUNT_ROUTE);
                 }}
               >
                 {t('importAccount')}
               </ButtonLink>
             </Box>
-            <Box>
+            <Box marginBottom={4}>
               <ButtonLink
                 size={Size.SM}
                 startIconName={IconName.Hardware}
@@ -257,71 +247,34 @@ export const AccountListMenu = ({ onClose }) => {
               >
                 {t('hardwareWallet')}
               </ButtonLink>
-              {
-                ///: BEGIN:ONLY_INCLUDE_IN(build-mmi)
-                <>
-                  <ButtonLink
-                    size={Size.SM}
-                    startIconName={IconName.Custody}
-                    onClick={() => {
-                      dispatch(toggleAccountMenu());
-                      trackEvent({
-                        category: MetaMetricsEventCategory.Navigation,
-                        event:
-                          MetaMetricsEventName.UserClickedConnectCustodialAccount,
-                      });
-                      if (getEnvironmentType() === ENVIRONMENT_TYPE_POPUP) {
-                        global.platform.openExtensionInBrowser(
-                          CUSTODY_ACCOUNT_ROUTE,
-                        );
-                      } else {
-                        history.push(CUSTODY_ACCOUNT_ROUTE);
-                      }
-                    }}
-                  >
-                    {t('connectCustodialAccountMenu')}
-                  </ButtonLink>
-                  {mmiPortfolioEnabled && (
-                    <ButtonLink
-                      size={Size.SM}
-                      startIconName={IconName.MmmiPortfolioDashboard}
-                      onClick={() => {
-                        dispatch(toggleAccountMenu());
-                        trackEvent({
-                          category: MetaMetricsEventCategory.Navigation,
-                          event:
-                            MetaMetricsEventName.UserClickedPortfolioButton,
-                        });
-                        window.open(mmiPortfolioUrl, '_blank');
-                      }}
-                    >
-                      {t('portfolioDashboard')}
-                    </ButtonLink>
-                  )}
-                  <ButtonLink
-                    size={Size.SM}
-                    startIconName={IconName.Compliance}
-                    onClick={() => {
-                      dispatch(toggleAccountMenu());
-                      trackEvent({
-                        category: MetaMetricsEventCategory.Navigation,
-                        event: MetaMetricsEventName.UserClickedCompliance,
-                      });
-                      if (getEnvironmentType() === ENVIRONMENT_TYPE_POPUP) {
-                        global.platform.openExtensionInBrowser(
-                          COMPLIANCE_FEATURE_ROUTE,
-                        );
-                      } else {
-                        history.push(COMPLIANCE_FEATURE_ROUTE);
-                      }
-                    }}
-                  >
-                    {t('compliance')}
-                  </ButtonLink>
-                </>
-                ///: END:ONLY_INCLUDE_IN
-              }
             </Box>
+            {
+              ///: BEGIN:ONLY_INCLUDE_IN(build-mmi)
+              <Box>
+                <ButtonLink
+                  size={Size.SM}
+                  startIconName={IconName.Custody}
+                  onClick={() => {
+                    dispatch(toggleAccountMenu());
+                    trackEvent({
+                      category: MetaMetricsEventCategory.Navigation,
+                      event:
+                        MetaMetricsEventName.UserClickedConnectCustodialAccount,
+                    });
+                    if (getEnvironmentType() === ENVIRONMENT_TYPE_POPUP) {
+                      global.platform.openExtensionInBrowser(
+                        CUSTODY_ACCOUNT_ROUTE,
+                      );
+                    } else {
+                      history.push(CUSTODY_ACCOUNT_ROUTE);
+                    }
+                  }}
+                >
+                  {t('connectCustodialAccountMenu')}
+                </ButtonLink>
+              </Box>
+              ///: END:ONLY_INCLUDE_IN
+            }
           </Box>
         </Box>
       ) : null}
